+++ acceptance-tests/tests/src/test/java/org/hyperledger/besu/tests/web3j/privacy/OnChainPrivacyAcceptanceTestBase.java
@@ -46,7 +46,7 @@ public class OnChainPrivacyAcceptanceTestBase extends PrivacyAcceptanceTestBase 
   }
 
   /**
-   * Crete an onchain privacy group. The privacy group id will be randomly generated.
+   * Create an onchain privacy group. The privacy group id will be randomly generated.
    *
    * <p>This method also checks that each node member has successfully processed the transaction and
    * has the expected list of member for the group.