+++ consensus/clique/src/test/java/org/hyperledger/besu/consensus/clique/jsonrpc/methods/CliqueProposalsTest.java
@@ -38,7 +38,7 @@ public class CliqueProposalsTest extends AbstractVoteProposerMethodTest {
 
   @Before
   public void setup() {
-    method = new CliqueProposals(getVoteProposer());
+    method = new CliqueProposals(getValidatorProvider());
   }
 
   @Test