+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/core/PrivacyParameters.java
@@ -20,6 +20,7 @@ import org.hyperledger.besu.crypto.KeyPair;
 import org.hyperledger.besu.crypto.KeyPairUtil;
 import org.hyperledger.besu.enclave.Enclave;
 import org.hyperledger.besu.enclave.EnclaveFactory;
+import org.hyperledger.besu.ethereum.privacy.PrivateStateGenesisAllocator;
 import org.hyperledger.besu.ethereum.privacy.PrivateStateRootResolver;
 import org.hyperledger.besu.ethereum.privacy.PrivateWorldStateReader;
 import org.hyperledger.besu.ethereum.privacy.storage.PrivacyStorageProvider;
@@ -29,12 +30,14 @@ import org.hyperledger.besu.ethereum.worldstate.WorldStateArchive;
 import org.hyperledger.besu.ethereum.worldstate.WorldStatePreimageStorage;
 import org.hyperledger.besu.ethereum.worldstate.WorldStateStorage;
 import org.hyperledger.besu.plugin.services.PrivacyPluginService;
+import org.hyperledger.besu.plugin.services.privacy.PrivacyGroupGenesisProvider;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.nio.file.Path;
 import java.util.Base64;
+import java.util.Collections;
 import java.util.Optional;
 
 import com.google.common.annotations.VisibleForTesting;
@@ -202,6 +205,24 @@ public class PrivacyParameters {
 
   public PrivacyPluginService getPrivacyService() {
     return privacyPluginService;
+  }
+
+  public PrivateStateGenesisAllocator getPrivateStateGenesisAllocator() {
+    // Note: the order of plugin registration may cause issues here.
+    // This is why it's instantiated on get. It's needed in the privacy pre-compile constructors
+    // but privacy parameters is built before the plugin has had a chance to register a provider
+    // and have cli options instantiated
+    return new PrivateStateGenesisAllocator(
+        onchainPrivacyGroupsEnabled, createPrivateGenesisProvider());
+  }
+
+  private PrivacyGroupGenesisProvider createPrivateGenesisProvider() {
+    if (privacyPluginService != null
+        && privacyPluginService.getPrivacyGroupGenesisProvider() != null) {
+      return privacyPluginService.getPrivacyGroupGenesisProvider();
+    } else {
+      return (privacyGroupId, blockNumber) -> Collections::emptyList;
+    }
   }
 
   @Override