+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/core/fees/EIP1559Test.java
@@ -16,13 +16,16 @@ package org.hyperledger.besu.ethereum.core.fees;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.hyperledger.besu.ethereum.mainnet.feemarket.BaseFeeMarket;
+import org.hyperledger.besu.ethereum.mainnet.feemarket.FeeMarket;
+
 import org.junit.Test;
 
 public class EIP1559Test {
 
   private static final long FORK_BLOCK = 783L;
   private final EIP1559 eip1559 = new EIP1559(FORK_BLOCK);
-  private final FeeMarket feeMarket = FeeMarket.eip1559();
+  private final BaseFeeMarket feeMarket = FeeMarket.london();
   private static final long TARGET_GAS_USED = 10000000L;
 
   @Test