+++ ethereum/eth/src/test/java/org/hyperledger/besu/ethereum/eth/manager/task/WaitForPeersTaskTest.java
@@ -32,7 +32,7 @@ import org.junit.Test;
 public class WaitForPeersTaskTest {
   private EthProtocolManager ethProtocolManager;
   private EthContext ethContext;
-  private final MetricsSystem metricsSystem = new NoOpMetricsSystem();;
+  private final MetricsSystem metricsSystem = new NoOpMetricsSystem();
 
   @Before
   public void setupTest() {